+++ src/components/cards/PartnerItemCard.tsx
